Add bootloader option to 'Reboot' tile
This commit adds the bootloader option that was missing here since we
have the reboot, recovery and power off option. It only makes sense to
include this option as well.

- Got rid of the if statements that are just hard to read at a glance
and used case switches. This way we can see from a distance WTF is going
on and not have to squint our eyes at it.

Change-Id: I8e8515d0260a5c1679772f8db53141f67b3d22ce
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>